+++ src/kasperl/writer/_text_file.py
@@ -1,5 +1,6 @@
 import abc
 import argparse
+import logging
 import os
 from typing import List, Dict
 
@@ -125,7 +126,11 @@ class TextFileWriter(StreamWriter, InputBasedPlaceholderSupporter, abc.ABC):
             item_str = self._data_formatter.format_data(item)
             item_str = self.session.expand_placeholders(item_str)
             output_file = self.session.expand_placeholders(self.output_file)
+
             self.logger().info("Writing to: %s" % output_file)
+            if self.logger().isEnabledFor(logging.DEBUG):
+                self.logger().debug(item_str)
+
             if self.append:
                 with open(output_file, "a") as fp:
                     fp.write(item_str)